Fix some compiler complaints

+++ include/reflection/base.hpp
@@ -80,7 +80,9 @@ struct UUID_t {
     bool operator == (const UUID_t& other) const { return memcmp(uuid, other.uuid, sizeof(uuid)) == 0; }
 
     int toString(char buffer[37]) const {
-        return sprintf(buffer, "%08x-%04x-%04x-%04x-%04x%08x", uuid[0], uuid[1] >> 16, uuid[1] & 0xffff, uuid[2] >> 16, uuid[2] & 0xffff, uuid[3]);
+        return sprintf(buffer, "%08x-%04x-%04x-%04x-%04x%08x",
+                (unsigned int)(uuid[0]), (unsigned int)(uuid[1] >> 16), (unsigned int)(uuid[1] & 0xffff),
+                (unsigned int)(uuid[2] >> 16), (unsigned int)(uuid[2] & 0xffff), (unsigned int)(uuid[3]));
     }
 };
